Proje 7 delete current kontrolu

+++ src/main/java/tr/com/metix/testproject/web/rest/CustomerResource.java
@@ -6,11 +6,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import tr.com.metix.testproject.domain.Customer;
 import tr.com.metix.testproject.domain.User;
+import tr.com.metix.testproject.repository.CustomerRepository;
 import tr.com.metix.testproject.security.SecurityUtils;
 import tr.com.metix.testproject.service.CustomerService;
 import tr.com.metix.testproject.service.UserService;
 import tr.com.metix.testproject.service.dto.CustomerDTO;
-import tr.com.metix.testproject.service.dto.UserDTO;
 import tr.com.metix.testproject.web.rest.errors.BadRequestAlertException;
 
 import javax.validation.Valid;
@@ -24,14 +24,18 @@ public class CustomerResource {
 
     private final CustomerService customerService;
     private final UserService userService;
+    private final CustomerRepository customerRepository;
 
     @Value("${jhipster.clientApp.name}")
     private String applicationName;
 
 
-    public CustomerResource(CustomerService customerService, UserService userService) {
+
+
+    public CustomerResource(CustomerService customerService, UserService userService, CustomerRepository customerRepository) {
         this.customerService = customerService;
         this.userService = userService;
+        this.customerRepository = customerRepository;
     }
 
 
@@ -40,7 +44,23 @@ public class CustomerResource {
     // /
     @DeleteMapping("/deletecustomer/{id}")
     public void deleteCustomer(@PathVariable Long id) {
-        customerService.deleteCustomer(id);
+
+        Optional<User> u = userService.getUserWithAuthoritiesByLogin(SecurityUtils.getCurrentUserLogin().get());
+
+
+        Optional<Customer> customer = customerRepository.findById(id);
+
+        System.out.println("cccccc : " + customer.get().getOwner().getId());
+
+        if(customer.get().getOwner().getId() == u.get().getId()){
+            System.out.println(" GİRİYOR : / " + u.get().getId() + " getownerId " + customer.get().getOwner().getId());
+            customerService.deleteCustomer(id);
+        }
+        else{
+            throw new BadRequestAlertException("Bu müşteriyi silemezsiniz", null, "test");
+        }
+
+
 //        return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, null, id.toString())).build();
     }
 
@@ -66,7 +86,7 @@ public class CustomerResource {
     @GetMapping("/customersupdate")
     public List<Long> getAllUsersUpdate(@RequestParam Long customerId) { // 4
 
-       Optional<User> u = userService.getUserWithAuthoritiesByLogin(SecurityUtils.getCurrentUserLogin().get());
+        Optional<User> u = userService.getUserWithAuthoritiesByLogin(SecurityUtils.getCurrentUserLogin().get());
 
         System.out.println("CURRENT USER : " + u.get().getId());
